Frequency tests [skip ci]
